Fix: ALL

+++ acted_monoid/min_add.hpp
@@ -1,11 +1,12 @@
 #pragma once
+#include <cstdint>
 #include <algorithm>
 
 struct MinAdd {
-	using S = long long;
-	using F = long long;
+	using S = int64_t;
+	using F = int64_t;
 	static S op(S a, S b) { return std::min(a, b); }
-	static S e() { return (long long)1e18; }
+	static S e() { return (int64_t)1e18; }
 	static S mapping(F f, S x) { return x + f; }
 	static F composition(F f, F g) { return f + g; }
 	static F id() { return 0; }
