+++ acted_monoid/min_affine.hpp
@@ -1,13 +1,14 @@
 #pragma once
+#include <cstdint>
 #include <algorithm>
 
 struct MinAffine {
-	using S = long long;
+	using S = int64_t;
 	struct F {
-		long long a, b;
+		int64_t a, b;
 	};
 	static S op(S a, S b) { return std::min(a, b); }
-	static S e() { return (long long)1e18; }
+	static S e() { return (int64_t)1e18; }
 	static S mapping(F f, S x) { return f.a * x + f.b; }
 	static F composition(F f, F g) { return {f.a * g.a, f.a * g.b + f.b}; }
 	static F id() { return {1, 0}; }
