+++ acted_monoid/sum_assign.hpp
@@ -1,12 +1,13 @@
 #pragma once
+#include <cstdint>
 #include <climits>
 
 struct SumAssign {
 	struct S {
-		long long val;
+		int64_t val;
 		int cnt;
 	};
-	using F = long long;
+	using F = int64_t;
 	static constexpr F NONE = LLONG_MIN;
 	static S op(S a, S b) { return {a.val + b.val, a.cnt + b.cnt}; }
 	static S e() { return {0, 0}; }
