+++ monoid/gcd.hpp
@@ -1,8 +1,9 @@
 #pragma once
+#include <cstdint>
 #include <numeric>
 
 struct GcdMonoid {
-	using value_type = long long;
+	using value_type = int64_t;
 	static value_type e() { return 0; }
 	static value_type op(value_type a, value_type b) { return std::gcd(a, b); }
 };
