+++ monoid/max.hpp
@@ -1,8 +1,9 @@
 #pragma once
+#include <cstdint>
 #include <algorithm>
 
 struct MaxMonoid {
-	using value_type = long long;
-	static value_type e() { return -(long long)1e18; }
+	using value_type = int64_t;
+	static value_type e() { return -(int64_t)1e18; }
 	static value_type op(value_type a, value_type b) { return std::max(a, b); }
 };
